sns_auto_poster/requirements.txt
google-genai>=1.0.0
tweepy>=4.14.0
requests>=2.31.0
pytz>=2024.1
pytrends>=4.9.2
Pillow>=10.0.0
